Measure with UserTiming and emit CSV when finished
Previously, this benchmark emitted its results per iteration to the console with
`console.time` and `console.timeEnd`. This is a bummer if you want to do any
kind of analysis on the data.

Instead, use `performance.mark` and `measure` to capture the timing data, and
then emit the results to the console in CSV format when finished.